refactor livechat
