+++ src/javascript/_common/utility.js
@@ -239,6 +239,9 @@ class PromiseClass {
     }
 }
 
+const lc_licenseID = 12049137;
+const lc_clientID = '66aa088aad5a414484c1fd1fa8a5ace7';
+
 module.exports = {
     showLoadingImage,
     getHighestZIndex,
@@ -259,4 +262,6 @@ module.exports = {
     getStaticHash,
     PromiseClass,
     removeObjProperties,
+    lc_licenseID,
+    lc_clientID,
 };
