+++ src/javascript/app/base/client.js
@@ -1,16 +1,18 @@
+const { init }           = require('@livechat/customer-sdk');
 const BinarySocket       = require('./socket');
 const Defaults           = require('../pages/trade/defaults');
 const RealityCheckData   = require('../pages/user/reality_check/reality_check.data');
 const ClientBase         = require('../../_common/base/client_base');
 const GTM                = require('../../_common/base/gtm');
 const SocketCache        = require('../../_common/base/socket_cache');
-const LiveChat           = require('../../_common/base/livechat');
 const { isBinaryDomain } = require('../../_common/utility');
 const getElementById     = require('../../_common/common_functions').getElementById;
 const removeCookies      = require('../../_common/storage').removeCookies;
 const urlFor             = require('../../_common/url').urlFor;
 const applyToAllElements = require('../../_common/utility').applyToAllElements;
 const getPropertyValue   = require('../../_common/utility').getPropertyValue;
+const licenseID          = require('../../_common/utility').lc_licenseID;
+const clientID           = require('../../_common/utility').lc_clientID;
 
 const Client = (() => {
     
@@ -86,6 +88,33 @@ const Client = (() => {
         }
     };
 
+    // Called when logging out to end ongoing chats if there is any
+    const endLiveChat = () => new Promise ((resolve) => {
+        const session_variables = { loginid: '', landing_company_shortcode: '', currency: '', residence: '', email: '' };
+        window.LiveChatWidget.call('set_session_variables', session_variables);
+        window.LiveChatWidget.call('set_customer_email', ' ');
+        window.LiveChatWidget.call('set_customer_name', ' ');
+        
+        try {
+            const customerSDK = init({
+                licenseId: licenseID,
+                clientId : clientID,
+            });
+            customerSDK.on('connected', () => {
+                if (window.LiveChatWidget.get('chat_data')) {
+                    const { chatId, threadId } = window.LiveChatWidget.get('chat_data');
+                    if (threadId) {
+                        customerSDK.deactivateChat({ chatId }).catch(() => null);
+                    }
+                }
+                resolve();
+            });
+        } catch (e){
+            resolve();
+        }
+
+    });
+
     const doLogout = (response) => {
 
         if (response.logout !== 1) return;
@@ -101,13 +130,12 @@ const Client = (() => {
         SocketCache.clear();
         RealityCheckData.clear();
         if (isBinaryDomain()){
-            LiveChat.endLiveChat().then(() => {
+            endLiveChat().then(() => {
                 redirection(response);
             });
         } else {
             redirection(response);
         }
-        
     };
 
     const getUpgradeInfo = () => {
